Fix structured page service
